requirements.txt
torch==2.1.0
torchaudio==2.1.0
transformers==4.36.2
gradio==3.50.2
openai-whisper==20231117